pyproject.toml
[project]
name = "movie-night-web"
version = "0.1.0"
description = "Web UI for movie night roll application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-api-core>=2.19.2",
    "google-api-python-client>=2.143.0",
    "google-auth-oauthlib>=1.2.1",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
